refactor: group interaction panels under panels subdirectory
