+++ frontend/src/components/pet/interactions/panels/battle/battle-result-art.tsx
