+++ frontend/src/components/pet/interactions/panels/battle/index.css
