+++ frontend/src/components/pet/interactions/panels/battle/index.tsx
