+++ frontend/src/components/pet/interactions/panels/battle/types.ts
