+++ frontend/src/components/pet/interactions/panels/level-up/index.tsx
