+++ frontend/src/components/pet/interactions/panels/rename/index.tsx
