+++ frontend/src/router/routes/battle-route/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import InteractionStandalone from '@components/pet/interactions/standalone';
-import BattlePanel from '@components/pet/interactions/battle-panel';
+import BattlePanel from '@components/pet/interactions/panels/battle';
 
 const BattleRoute: React.FC = () => (
     <InteractionStandalone action="battle" minPets={1}>
