+++ frontend/src/router/routes/breed-route/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import InteractionStandalone from '@components/pet/interactions/standalone';
-import BreedPanel from '@components/pet/interactions/breed-panel';
+import BreedPanel from '@components/pet/interactions/panels/breed';
 
 /** Top-level `/breed` — shell + breed panel (standalone UI). */
 const BreedRoute: React.FC = () => (
