+++ frontend/src/router/routes/level-up-route/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import InteractionStandalone from '@components/pet/interactions/standalone';
-import LevelUpPanel from '@components/pet/interactions/level-up-panel';
+import LevelUpPanel from '@components/pet/interactions/panels/level-up';
 
 const LevelUpRoute: React.FC = () => (
     <InteractionStandalone action="levelup" minPets={1}>
