+++ frontend/src/router/routes/rename-route/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import InteractionStandalone from '@components/pet/interactions/standalone';
-import RenamePanel from '@components/pet/interactions/rename-panel';
+import RenamePanel from '@components/pet/interactions/panels/rename';
 
 const RenameRoute: React.FC = () => (
     <InteractionStandalone action="changename" minPets={1}>
